Remove foreign as a separate module to simplify.
